+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Changelog
 
 ### unreleased
-+ Mock single requests
-+ Add delay to mocked responses
++ Mock single requests to NSURLConnection
++ Add delay to mocked responses on NSURLConnection
++ Mock NSURLSession simple init methods